Modify the type of input from int to uint16 because romans numeral can't go more than 3999

+++ Property-Based-Test/numeral_bench_test.go
@@ -7,7 +7,7 @@ func BenchmarkArabicToRomans(b *testing.B) {
 	for _, c := range cases {
 		b.Run("arabics to romans", func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
-				ConvertToRoman(c)
+				ConvertToRoman(uint16(c))
 			}
 		})
 	}
